[C++Seminar] define all in header file
